+++ src/main/kotlin/minesweeper/domain/cell/CellType.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper.domain.cell
 
 enum class CellType {
     NONE,
