+++ src/main/kotlin/minesweeper/domain/cell/Column.kt
@@ -1,4 +1,4 @@
-package minesweeper.domain
+package minesweeper.domain.cell
 
 @JvmInline
 value class Column(
